New version 0.5.2.0 created.
New features / improvements:
- None

Removed features:
- None

Fixed bugs:
- The language entries for the brokerage labels have been added.
- The bug in the edit of a share has been fixed. It was not possible to save a changed dividend
  because the check if the document of the dividend is already used. This check has been corrected.
- In the sale overview the loss was as "-" and not the value itself. This incorrect show has been fixed.
- If in the add share for a invalid web address for the daily values is given the cursor jumps to the wrong text box. This has been fixed.
- The costs are shown now correct.
- The values "Interval" and "Amount" in the Settings.xml for the charting is now used by the software.
- The incorrect document parsing of the direct document parsing works now.

+++ SharePortfolioManager/Forms/ShareDetailsForm/FormFiles/TabControl.cs
@@ -91,7 +91,7 @@ namespace SharePortfolioManager.ShareDetailsForm
                                 LanguageName) +
                             @" " +
                             Language.GetLanguageTextListByXPath(@"/ComboBoxItemsShareType/*", LanguageName)[
-                                ShareObjectMarketValue.ShareType];
+                                (int)ShareObjectMarketValue.ShareType];
                     }
                     else
                     {
@@ -105,7 +105,7 @@ namespace SharePortfolioManager.ShareDetailsForm
                                 LanguageName) +
                             @" " +
                             Language.GetLanguageTextListByXPath(@"/ComboBoxItemsShareType/*", LanguageName)[
-                                ShareObjectMarketValue.ShareType];
+                                (int)ShareObjectMarketValue.ShareType];
                     }
 
                     if (ShareObjectMarketValue.LastUpdateShare == DateTime.MinValue)
@@ -268,7 +268,7 @@ namespace SharePortfolioManager.ShareDetailsForm
                                 LanguageName) +
                             @" " +
                             Language.GetLanguageTextListByXPath(@"/ComboBoxItemsShareType/*", LanguageName)[
-                                ShareObjectFinalValue.ShareType];
+                                (int)ShareObjectFinalValue.ShareType];
                     }
                     else
                     {
@@ -282,7 +282,7 @@ namespace SharePortfolioManager.ShareDetailsForm
                                 LanguageName) +
                             @" " +
                             Language.GetLanguageTextListByXPath(@"/ComboBoxItemsShareType/*", LanguageName)[
-                                ShareObjectFinalValue.ShareType];
+                                (int)ShareObjectFinalValue.ShareType];
                     }
 
                     if (ShareObjectFinalValue.LastUpdateShare == DateTime.MinValue)
